Clear out ml2_vlan_allocations if CVX has no VLAN pool configured
If there is no vlan pool configured for a region, but CVX was
reachable, the arista_vlan type driver would fail to start.
We should treat this case as there being no VLANs assigned
to the region and clear out all unallocated vlans from the db.

+++ networking_arista/ml2/drivers/driver_helpers.py
@@ -125,6 +125,9 @@ class VlanSyncService(object):
                     'update'))
 
             for alloc in allocs:
+                if alloc.allocated:
+                   continue
+
                 if alloc.physical_network != 'default':
                     session.delete(alloc)
